splitting out settings, environment, and browser code into separate modules
