+++ tests/integration/hooks.js
@@ -1,10 +1,10 @@
-const utils = require('../utils');
+const environment = require('./utils/environment');
 const request = require('request-promise-native');
 
 exports.mochaHooks = {
   beforeAll: async ()=> {
     console.log('Starting services......');
-    await utils.prepServices();
+    await environment.prepServices();
     console.log('Services started');
   },
