+++ tests/utils/browser.js
@@ -1,4 +1,6 @@
 const fs = require('fs');
+let browserLogStream;
+
 const feedBackDocs = async (testName = 'allLogs', existingDocIds = []) => {
   const feedBackDocs = await browser.executeAsync(feedBackDocsScript);
   const flattened = feedBackDocs.flat();
@@ -28,7 +30,28 @@ const getCookies = (...cookieNameList) => {
   return browser.getCookies(cookieNameList);
 };
 
+const saveBrowserLogs = () => {
+  // wdio also writes in this file
+  if (!browserLogStream) {
+    browserLogStream = fs.createWriteStream(__dirname + '/../logs/browser.console.log');
+  }
+
+  return browser
+    .manage()
+    .logs()
+    .get('browser')
+    .then(logs => {
+      const currentSpec = jasmine.currentSpec.fullName;
+      browserLogStream.write(`\n~~~~~~~~~~~ ${currentSpec} ~~~~~~~~~~~~~~~~~~~~~\n\n`);
+      logs
+        .map(log => `[${log.level.name_}] ${log.message}\n`)
+        .forEach(log => browserLogStream.write(log));
+      browserLogStream.write('\n~~~~~~~~~~~~~~~~~~~~~\n\n');
+    });
+};
+
 module.exports = {
   feedBackDocs,
-  getCookies
+  getCookies,
+  saveBrowserLogs
 };
