+++ tests/utils/environment.js
@@ -1,6 +1,32 @@
 const utils = require('../utils');
 const rpn = require('request-promise-native');
-const constants = require('./constants');
+const constants = require('../constants');
+const settings = require('./settings');
+
+const runAndLogApiStartupMessage = (msg, func) => {
+  console.log(`API startup: ${msg}`);
+  return func();
+};
+
+const apiRetry = () => {
+  return new Promise(resolve => {
+    setTimeout(() => {
+      resolve(listenForApi());
+    }, 1000);
+  });
+};
+
+const listenForApi = async () => {
+  console.log('Checking API');
+  try {
+    await utils.request({ path: '/api/info' });
+    console.log('API is up');
+  } catch (err) {
+    console.log('API check failed, trying again in 1 second');
+    console.log(err.message);
+    await apiRetry();
+  }
+};
 
 const prepServices = async (config) => {
   if (constants.IS_CI) {
@@ -9,7 +35,7 @@ const prepServices = async (config) => {
     // getting pushed into horti.log Once horti has bootstrapped we want to restart everything so
     // that the service processes get restarted with their logs separated and pointing to the
     // correct logs for testing
-    await utils.listenForApi();
+    await listenForApi();
     console.log('Horti booted API, rebooting under our logging structure');
     await rpn.post('http://localhost:31337/all/restart');
   } else {
@@ -17,25 +43,20 @@ const prepServices = async (config) => {
     await rpn.post('http://localhost:31337/all/start');
   }
 
-  await utils.listenForApi();
+  await listenForApi();
   if (config && config.suite === 'web') {
-    await utils.runAndLogApiStartupMessage('Settings setup', setupSettings);
+    await runAndLogApiStartupMessage('Settings setup', settings.setupSettings);
   }
-  await utils.runAndLogApiStartupMessage('User contact doc setup', utils.setUserContactDoc);
+  await runAndLogApiStartupMessage('User contact doc setup', utils.setUserContactDoc);
 };
 
-const setupSettings = () => {
-  const defaultAppSettings = utils.getDefaultSettings();
-  defaultAppSettings.transitions = {};
-
-  return utils.request({
-    path: '/api/v1/settings?replace=1',
-    method: 'PUT',
-    body: defaultAppSettings
-  });
+const tearDownServices = () => {
+  return rpn.post('http://localhost:31337/die');
 };
 
 
 module.exports = {
-  prepServices
+  prepServices,
+  tearDownServices,
+  runAndLogApiStartupMessage
 };
