+++ tests/wdio.conf.js
@@ -4,7 +4,7 @@ const fs = require('fs');
 const constants = require('./constants');
 const utils = require('./utils');
 const path = require('path');
-
+const environment = require('./utils/environment');
 const chai = require('chai');
 chai.use(require('chai-exclude'));
 
@@ -200,7 +200,7 @@ const baseConfig = {
     if (fs.existsSync(browserLogPath)) {
       fs.unlinkSync(browserLogPath);
     }
-    await utils.prepServices(config);
+    await environment.prepServices(config);
   },
   /**
    * Gets executed before a worker process is spawned and can be used to initialise specific service
